+++ tests/test_request.py
@@ -1,0 +1,32 @@
+def test_wait_between_retries():
+    # first try at t=0, then t=5, then t=15
+    pass
+
+
+def test_timeout_retry_count_2():
+    # should make 3 attempts total when retry = 2
+    pass
+
+
+def test_timeout_retry_count_0():
+    # should make 1 attempt when retry = 0
+    pass
+
+
+def test_timeout_retry_count_negative():
+    # should make 1 attempt when retry less than 0
+    # print warning message
+    pass
+
+
+def test_network_error():
+    # aiohttp.ClientError
+    pass
+
+
+def test_html_response():
+    pass
+
+
+def test_malformed_json_response():
+    pass
